+++ CHANGELOG.md
@@ -7,3 +7,8 @@ Initial release
 ## [1.0.1] - 2026-02-25
 
 Enable http errors
+
+## [1.1.0] - 2026-06-08
+
+Add response callback
+Disable http errors